Improve error handling and add tests
Error handling had not properly been setup in the package. Errors
from the primary function would largely be handled by a "panic"
which is not great for callers downstream.

Errors are now properly returned to the caller to deal with
themselves. As a result testing can also be done more intentionally.

Tests have been added to try out various different scenarios but
largely to make sure the output of the package is as predictable as
possible. Table-driven testing has been used. There is likely some
refactoring to do and plenty of more tests can be added.

As the interface changes this will up the major version of the
package.

The reference implementation for the library has been updated.