Add "Human" tests to enable its removal
Part of ross-spencer/spargo#4 tests are added to compare what is
added to "Human" in the results structure. The tests help to prove
that it should be possible to serialize the results structure as-is
to a string and it will match what we had previously. Part of a 2.x.x
release to tackle this issue.

The tests helped to identify other missing fields and they have been
added as a result. It also helped to identify what users may expect
from SPARQL results when serialized to JSON and so makes those more
idiomatic. JSON tags are correct now and should export lower case
datatypes.

Tests will be removed in the following commit with the rest of
the code using the "Human" field.