Remove .Human from SPARQLResult
The field used here should never have made it to the original code
but it did. We have tested the equivalence to the serialization to
JSON of the structures without it and so all should be good for the
next release. This will up the major version by one again.

Resolves ross-spencer/spargo#4